curriculum-datasetGen/requirements.txt
pyyaml>=6.0
transformers>=4.35.0
torch>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
GitPython>=3.1.40
pathspec>=0.11.0
